fix: 친구검색 시 API를 호출하는 시점에 따라 undefined을 반환받는 현상 수정 #150
컴포넌트를 분리하여 isLoading, isError 분기처리 추가
